+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Read size underflow on read error.
 - Declaration and usage of minBufferSize.
 
+### Changed
+- Remaining static const members to constexpr.
+
 ## [1.1.0] - 2024-08-29
 ### Changed
 - Marked UartServer class as deprecated (StreamServer should be used instead).